revert: split api
This commit reverts the split API implementation.

The crate now simply provides a sync `detect`
method and `subscribe` remains async, also
removed `notify` files.
