+++ examples/subscribe.rs
@@ -1,13 +1,9 @@
-use std::error::Error;
-
 use futures_lite::StreamExt;
 
 #[tokio::main]
-async fn main() -> Result<(), Box<dyn Error>> {
+async fn main() {
     let mut stream = dark_light::subscribe().await;
     while let Some(mode) = stream.next().await {
-        println!("System theme changed: {:?}", mode);
+        println!("System mode changed: {:?}", mode);
     }
-
-    Ok(())
 }
