+++ src/platforms/freedesktop/detect.rs
@@ -1,14 +1,3 @@
-#[cfg(any(feature = "sync", doc))]
-pub mod sync {
-    use crate::Mode;
-
-    pub fn detect() -> Mode {
-        futures_lite::future::block_on(super::super::get_color_scheme())
-    }
-}
-
-use crate::Mode;
-
-pub async fn detect() -> Mode {
-    super::get_color_scheme().await
+pub fn detect() -> crate::Mode {
+    futures_lite::future::block_on(super::get_color_scheme())
 }
