+++ src/platforms/websys/detect.rs
@@ -1,4 +1,4 @@
-pub fn detect_mode() -> crate::Mode {
+pub fn detect() -> crate::Mode {
     if let Some(window) = web_sys::window() {
         let query_result = window.match_media("(prefers-color-scheme: dark)");
         if let Ok(Some(mql)) = query_result {
@@ -6,15 +6,4 @@ pub fn detect_mode() -> crate::Mode {
         }
     }
     crate::Mode::Light
-}
-
-#[cfg(any(feature = "sync", doc))]
-pub mod sync {
-    pub fn detect() -> crate::Mode {
-        detect_mode()
-    }
-}
-
-pub async fn detect() -> crate::Mode {
-    detect_mode()
 }
