+++ src/platforms/websys/subscribe.rs
@@ -1,9 +1,9 @@
 pub async fn subscribe() -> impl futures_lite::Stream<Item = crate::Mode> {
     Box::pin(futures_lite::stream::unfold(
-        crate::detect().await,
+        crate::detect(),
         |last_mode| async move {
             loop {
-                let current_mode = crate::detect().await;
+                let current_mode = crate::detect();
 
                 if current_mode != last_mode {
                     return Some((current_mode, current_mode));
